Handle keyboard method visible.
SystemListItem: Fix null property warnings.  On edit text focus update
last known index, so the view may scroll to it when keyboard is shown.
Focus did not always work, so activeFocus will be used.
Main: Add property lastFocus, which is the item to scroll to when
soft keyboard is shown.  When soft keyboard is show we want to hide
the header, resize the list to visible area, and scroll to
the currently editing item.
Add icon to page header, for fun gooey.
Remove theme name, which is not working properly.
Uncheck chkSelectAll when items are removed.