Move command output to a page
Instead of opening a popup, move to the output page.
Main: Add Output to pages, and expose signal to move to it.
Output: Popover => Page.  To Bind to a possibly undefined buffer,
we have to use a Loader that binds only when visible.
ProcessItem: Expose process buffer to move to a page instead of
a popup.
ScriptPage: Expose process buffer to move to a page instead of a popup.